Reduce sizes

+++ PKHeX.Core/Saves/SAV3.cs
@@ -24,7 +24,7 @@ namespace PKHeX.Core
         private const int SIZE_RESERVED = 0x10000; // unpacked box data will start after the save data
         public const int SIZE_BLOCK_USED = 0xF80;
 
-        private static readonly int[] chunkLength =
+        private static readonly ushort[] chunkLength =
         {
             0xf2c, // 0 | Small Block (Trainer Info)
             0xf80, // 1 | Large Block Part 1
